nexuscore/backend/requirements/development.txt
-r base.txt

# Development Tools
django-debug-toolbar>=4.3.0
django-silk>=5.0.0
django-querycount>=0.8.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
pytest-mock>=3.12.0
factory-boy>=3.3.0
faker>=23.0.0
freezegun>=1.3.0
responses>=0.25.0

# Code Quality
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
flake8-django>=1.4.0
mypy>=1.8.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
sphinxcontrib-django>=2.3.0

# Security Scanning
bandit>=1.7.7
safety>=3.0.0